+++ dwds/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 12.0.0-dev
+
+- Implement `lookupResolvedPackageUris` and `lookupPackageUris` vm service API.
+- Update `vm_service` version to `^7.5.0`.
+- Make `ExpressionCompilerService` infer location of `libraries.json` from
+  `sdkDir` parameter.
+- Show an alert in the Dart Debug Extension for a multi-app scenario.
+- Fix a bug where `dartEmitDebugEvents` was set as a `String` instead of `bool`
+  in the injected client.
+
+**Breaking changes:**
+
+- Add `sdkDir` and `librariesPath` arguments to `Dwds.start` to help file
+  resolution for sdk uris.
+- Add `emitDebugEvents` argument to `Dwds.start` to suppress emitting debug
+  events from the injected client.
+- Replace `sdkRoot` parameter by `sdkDir` in `ExpressionCompilerService`.
+
+## 11.5.1
+
+- Update SDK contraint to `>=2.15.0 <3.0.0`.
+
 ## 11.5.0
 
 - Support hot restart in a multi-app scenario with legacy modules.
@@ -14,7 +36,7 @@
 - Fix chrome detection in iPhone emulation mode in chrome or edge browsers.
 - Reliably find unused port for extension backend http service.
 - Ignore offset / count parameters in getObject if the object has no length.
-- Include static member information for classes
+- Include static member information for classes.
 
 ## 11.4.0